Implementation of PC2
